+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/eventpublishing/LinkedEventPublisher.java
@@ -3,7 +3,9 @@ package uk.gov.justice.services.eventsourcing.eventpublishing;
 import static java.lang.String.format;
 import static javax.transaction.Transactional.TxType.REQUIRES_NEW;
 
+import uk.gov.justice.services.eventsourcing.eventpublishing.configuration.EventPublishingWorkerConfig;
 import uk.gov.justice.services.eventsourcing.publishedevent.EventPublishingException;
+import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.CompatibilityModePublishedEventRepository;
 import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.EventPublishingRepository;
 import uk.gov.justice.services.eventsourcing.publisher.jms.EventPublisher;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.LinkedEvent;
@@ -26,18 +28,33 @@ public class LinkedEventPublisher {
     @Inject
     private LinkedJsonEnvelopeCreator linkedJsonEnvelopeCreator;
 
+    @Inject
+    private EventPublishingWorkerConfig eventPublishingWorkerConfig;
+
+    @Inject
+    private CompatibilityModePublishedEventRepository compatibilityModePublishedEventRepository;
+
     @Transactional(REQUIRES_NEW)
     public boolean publishNextNewEvent() {
 
         final Optional<UUID> eventId = eventPublishingRepository.popNextEventIdFromPublishQueue();
         if (eventId.isPresent()) {
-            final Optional<LinkedEvent> linkedEvent = eventPublishingRepository.findEventFromEventLog(eventId.get());
+            final Optional<LinkedEvent> linkedEventOptional = eventPublishingRepository.findEventFromEventLog(eventId.get());
 
-            if (linkedEvent.isPresent()) {
-                final JsonEnvelope linkedJsonEnvelope = linkedJsonEnvelopeCreator.createLinkedJsonEnvelopeFrom(linkedEvent.get());
+            if (linkedEventOptional.isPresent()) {
+                final LinkedEvent linkedEvent = linkedEventOptional.get();
+                final JsonEnvelope linkedJsonEnvelope = linkedJsonEnvelopeCreator.createLinkedJsonEnvelopeFrom(linkedEvent);
                 eventPublisher.publish(linkedJsonEnvelope);
                 eventPublishingRepository.setIsPublishedFlag(eventId.get(), true);
+
+                // Temporary. To be removed once the migration to the new publishing is released
+                // and published_event table is deleted
+                if(eventPublishingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable()) {
+                    compatibilityModePublishedEventRepository.insertIntoPublishedEvent(linkedJsonEnvelope);
+                }
+                
                 return true;
+
             } else {
                 throw new EventPublishingException(format("Failed to find LinkedEvent in event_log with id '%s' when id exists in publish_queue table", eventId.get()));
             }